pyproject.toml
[project]
name = "yearlyreps2"
version = "0.1.0"
description = "yearlyreps.tompaton.com"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.7",
    "django-ninja>=1.3.0",
    "mako>=1.3.9",
]
